requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gunicorn==23.0.0
huggingface-hub==0.30.2
icecream==2.1.4
idna==3.10
multidict==6.4.3
openai==0.28.1
packaging==25.0
paramiko==3.5.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.20.0
